third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/open-close/no_window_open_when_term_nesting_level_nonzero.window-expected.txt

This is a testharness.js-based test.
PASS no popups with frame removal
PASS no popups with frame navigation
PASS no popups from synchronously reachable window
FAIL no popups from another synchronously reachable window assert_equals: expected no popup during beforeunload expected null but got object "[object Window]"
Harness: the test ran to completion.
